move slug generation functionality into mixin.
